+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming version
 Bug fixes:
 - Fix missing download extension of converted text submission
+- Fault tolerance when no grade level is found for final points
 
 ## Version 3.4 (2024-11-04)
 Plugin: